+++ digital_asset_types/src/dao/scopes/asset.rs
@@ -7,7 +7,7 @@ use crate::{
         sea_orm_active_enums::V1AccountAttachments,
         token_accounts, tokens, FullAsset, Pagination, SearchAssetsQuery,
     },
-    rpc::{filter::TokenTypeClass, options::Options},
+    rpc::{filter::TokenType, options::Options},
 };
 use indexmap::IndexMap;
 use sea_orm::{
@@ -1275,12 +1275,12 @@ where
                         extensions::asset::Column::OwnerType,
                     )
                     .eq(OwnerType::Token),
-                    TokenType::NonFungible | TokenType::Nft => Expr::tbl(
+                    TokenType::NonFungible | TokenType::RegularNFT => Expr::tbl(
                         extensions::asset::Entity,
                         extensions::asset::Column::OwnerType,
                     )
                     .eq(OwnerType::Single),
-                    TokenType::Compressed => {
+                    TokenType::CompressedNFT => {
                         Expr::tbl(extensions::asset::Entity, extensions::asset::Column::TreeId)
                             .is_not_null()
                     }
